mv_rag_poc/requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langchain-ollama>=0.2.0
chromadb>=0.5.0
ollama>=0.3.0
networkx>=3.3
streamlit>=1.35.0
pydantic>=2.0.0
PyGithub>=2.3.0
python-dotenv>=1.0.0
mcp>=1.0.0
atlassian-python-api>=3.41.0
pypdf>=4.0.0